Backend/requirements.txt
fastapi
uvicorn[standard]
oracledb
pydantic
python-dotenv
PyJWT
